requirements.txt
django >= 3.0
psycopg2-binary
django-bootstrap-modal-forms
django-widget-tweaks